Internal: Change snapshot state for unreleased versions and add validation tests for constants
Currently the snapshot flag for Version constants is only set to true
for CURRENT.  However, this means that the snapshot state changes from
branch to branch.  Instead, snapshot should be "is this version
released?".  This change also adds a validation test checking that
ID -> constant and vice versa are correct, and fixes one bug found there
(for an unreleased version).


Former-commit-id: f2411253026503954da172ddd0b1bf77789dd476