Mapping: add an assertion to verify consistent serialization
We recently run into two issues where mapping weren't serialized in a consistent manner (#10302 and #10318). We rely on this consistency to do a byte level checl that mappings we get from the master are indentical to the one we have locally. Mistakes here can cause endless refresh mapping loops.

This commit adds an assert that verifies this upon every update from the master.


Former-commit-id: 521f804c7dcb3173efca2abbeabedb901b2bd2d7