Add prediction capability to MovAvgReducer
This commit adds the ability for moving average models to output a "prediction" based on the current
moving average model.  For simple, linear and single, this prediction is simply converges on the
moving average's mean at the last point, leading to a straight line.  For double, this will
predict in the direction of the linear trend (either globally or locally, depending on beta).

Also adds some more tests.

Closes #10545


Former-commit-id: 30177887b155a4bcb44cf1ec257fd1ae81028661