[ENGINE] Implement retries for ShadowEngine creation
When using a filesystem that may have lag between an index being created
on the primary and a on the replica, creation of the ShadowEngine can
fail because there are no segments in the directory.

In these situations, we retry during engine creation to wait until an
index is present in the directory. The number wait delay is
configurable, defaulting to waiting for 5 seconds from an index to
become available.

Resolves #10637


Former-commit-id: 2627324ac2399c43a59956d67856e02abec6f5d2