Internal: Remove the query parser cache.
The original goal of this cache was to avoid parsing the same query several
times in case several shards are held on the same node. While this might
sound like a good idea, this would only help when parsing the query takes
non-negligible time compared to actually running the query, which should not
be the case.


Former-commit-id: 9d890c472bfd328fb69ff07d98cfcdfe215e21b1