[ENGINE] Make translog fully private to the engine
This commit moves the translog creation into the InternalEngine
to ensure the transactino log is created after we acquired the write
lock on the index. This also prevents races when ShadowEngines are shutting
down due to node restarts where another node already takes over the not yet
fully synced transaction log.


Former-commit-id: 767309265b71f7fe9bb235c996662f2d70710a75