Close ShardFilterCache after Store is closed
The ShardFilterCache relies on the fact that it's
closed once the last reader on the shard is closed.
This is only guaranteed once the Store and all its
references are closed. This commit moves the closing
into the internal callback mechanism we use for deleting
shard data etc. to close the cache once we have all
searchers released.


Former-commit-id: 5cd6ced7eefaea10882ea510d9b1f473c7f45860