Make translog file name parsing strict
Today we are very lenient in parsing the translog files. This is
actually not necessary since we have a clear run once upgrade path.
All files are converted into the new file name pattern such that we
only need to look at old file patterns in the context of the upgrade.

This commit makes parsing really strict with the exceptoin of the upgrade path.


Former-commit-id: daca05b36efd486461ceb6ca1dfd92c54a10f540